refactor: Add platform-specific scroll scaling and async OS operations (#63)
- Refactor ComputerInterface to use ThreadPoolExecutor for all blocking OS calls
  - Prevents blocking the main asyncio event loop
  - Ensures low latency and seamless operation
- Add platform-specific scroll scaling (Windows: 120 units per tick, Mac/Linux: 1 unit)
  - Windows scroll clicks now properly scaled to WHEEL_DELTA (120) standard
  - Ensures scroll clicks match physical mouse wheel behavior
- Update documentation to reflect scroll behavior and async improvements
- Add test_scroll.py script for testing scroll functionality

+++ backend/src/services/vision/internvl.py
@@ -77,7 +77,7 @@ class InternVLModel:
             try:
                 self.model = AutoModel.from_pretrained(
                     self.model_name,
-                    torch_dtype=torch.bfloat16,
+                    dtype=torch.bfloat16,
                     low_cpu_mem_usage=True,
                     use_flash_attn=use_flash_attn,
                     device_map="auto",  # Let accelerate decide device placement
@@ -98,7 +98,7 @@ class InternVLModel:
                     self.model = (
                         AutoModel.from_pretrained(
                             self.model_name,
-                            torch_dtype=dtype,
+                            dtype=dtype,
                             low_cpu_mem_usage=True,
                             use_flash_attn=False,
                             trust_remote_code=self.trust_remote_code,
@@ -118,7 +118,7 @@ class InternVLModel:
                         self.model = (
                             AutoModel.from_pretrained(
                                 self.model_name,
-                                torch_dtype=torch.float32,
+                                dtype=torch.float32,
                                 low_cpu_mem_usage=True,
                                 use_flash_attn=False,
                                 trust_remote_code=self.trust_remote_code,
